PS interpreter - fix EPSFitPage (again)
Bug #700279 "On square EPS files -dFitPage incorrectly rotates output"

The commit to fix bug #698555 altered the action of EPSFitPage so that
it set the media size form the BoundingBox (using EPSCrop) and then
relied on the usual PostScript page fitting.

However, as noted at the time, this does cause square pages to rotate.
This isn't actually incorrect, there is no specification for this its a
Ghostscript feature, and we're free to implement it any way we like.
However it is a change form the previous behaviour.

This commit reinstates the EPSFitPage code (in EPSBoundingBoxFitPage
in gs_epsf.ps) and fixes the calculations in that code when the content
is rotated onto the media (because it fits better that way) so that it
correctly scales and translates the content onto the new media.

because we are no longer using the PostScript page fitting code, this
no longer rotates square media.