Bug 700278: Fix inverted colors in jpegcmyk output.
It seems that in the transform pixel region code, I know that
colours are in the right space, and therefore assume that there
will be a 1:1 mapping from the raw bytes into the device memory.

Sadly the jpegcmyk device inverts everything, so my assumption
was wrong.

Accordingly I've added a gxdso to tell me if the encoding is
direct. I've added this to the obvious devices - more may be
required in future.

If this is unimplemented, then we'll take the slower route
and look up colors.