Remove antiquated check in pdf_shift_text_currentpoint.
pdf_shift_text_currentpoint tests to see that the graphics state
it is passed is really a graphics state. We believe this is a
hangover from having graphics states and imager states.

Firstly, the check is wrong, as it uses penum->dev->memory when
it should be using penum->pgs->memory. Fixing this is enough to
solve the problem where this was seen (gpdl pdfwriting files).

Secondly, however the check is no longer needed, so the nicer fix
is to remove it completely.