pdfwrite - add a check to ensure we don't create an invalid PDF file
related to bug #700291 The PDF file in that bug report was created by
Ghostscript and pdfwrite. The xref table is invalid because it has an
entry with an offset of 11 digits, which is illegal (the file exceeds
10GB in size)

Add a check when writing the xref table to see if any offset is greater
that 10 digits, and throw an error if it is. We can't check any earlier
because we don't know the offset of any object in the file until it is
read from its stored temporary file and written to the main output
file.