GPDL: Run the pdfwrite setup each time we start a PS/PDF job.
Re-run the "scheduled inits" procs after installing the device
into the gstate - for the benefit of pdfwrite.

For neatness, move the scheduled inits list dictionary, and the proc
to run them into internaldict, so the don't contaminate the "normal"
name space in systemdict.

(Whatever git might say, the cleverness here is due to Chris).