pdfwrite - conform to Acrobat limitations for PDF 1.4 to conform to PDF/A-1
Never noticed before, but PDF/A-1 mandates sticking to the architectural
limits of Acrobat described in the PDF 1.4 specification (Note, these
are not *PDF* Limitations, they are limitations of the Adobe
implementation).

It was possible for us to emit co-ordinate movements which are in real
numbers, but exceed 32767, which is outside the Acrobat limitation.

This commit alters that, and in fact uses the older limitation of
Acrobat 3 and 4, which limits co-ordinates to +/- 16K.