pdfwrite - don't try and apply transfer functions when converting colours
Bug #700052 "pdfwrite device grayscale conversion looses significant quality on gradients

The problem is that the file contains considerable spurious (ie has no
effect) transparency operatiosns. in this particular case a number of
objects are drawn with an SMask and the SMask uses a transfer function
to invert the single component colour space (The gray objects are in
fact described in a DeviceN [/Black] colour space, presumabkly to
still further pointlessly increase teh complexity).

When converting colour space we were altering the transfer function
from 'Presever' to 'Apply', but we can't actually apply the transfer
function to the SMask in this case.

So leave this as 'preserve for now', note that transfer functions are
deprecated, and not permitted in PDF 2.0.