Add an "r_is_number()" macro
By moving the t_integer and t_real enum values to consecutive positions,
starting on a multiple of 2, we can check for a "number" type in a single step.

Useful for values specced as reals, but which may be integer values, and thus
parsed into integer objects. This means validation can be a single check,
rather than checking for real and integer individually.

Also use the macro in various places.