txtwrite - allow some slop in zero-width detection
Bug #700322 " 	Endless loop when converting pdf to text "

While trying to fit text (text format 2 or 3) we do some heuristic
juggling to decide whether or not to use a calculated minimum size.
During this we check to see whether the width is zero, and if it is
we don't try to use it because it would result in an infinite loop.

We actually need to allow a little 'slop' around zero, in case floating
point errors result in a very small, but not quite zero, width.

The loop resuling in the bug isn't endless, but it is very lengthy.