PDF interpreter - Improve T42 CharStrings creation for non-symbolic fonts
Bug #700158 "ghostscript produces incorrect JPG if SubsetFonts=false"

The problem is not in fact a problem with the pdfwrite device, it is
in fact the poorly documented way that Acrobat deals with TrueType
fonts, the symbolic/non-symbolic flags and Encodings.

When the font has a 3,1 CMAP subtable we use a number of sources of data
to try and build a CharStrings dictionary, which we use to map a glyph
name to a glyph ID:

               Encoding             CharStrings
character code --------> glyph name -----------> glyh ID

However when the font has no 3,1 CMAP subtable the specification says
we should try to use a 1,0 CMAP instead. For 1,0 tables we were not
using the 'prebuilt_encoding' to generate the CharStrins, which meant
that we making less use of information than we might, and some glyphs
were replaced by the /.notdef when rendering.

This commit adds the prebuilt_encoding to the sources of information
we use to generate Charstrings when using a 1,0 CMAP subtable.

This fixes the problem, and shows a progression with one file in our
test suite.