Fix header inclusions.
Run a release build on a linux machine to make arch.h etc. Then
run toolbin/headercompile.pl to test compiling each headerfile
by itself.

Resolve all the missing #includes, add missing repeated include
guards and copyright statements.

Also, update all the header dependencies in the makefiles.

It is possible that the object dependencies in the makefiles can be
simplified now, but that's a task for another day.