Change zmisc.c implementation of usertime to start at 0
The Display Postscript implementation in zcontext.c always returns 0
the first time 'usertime' is called, and counts time from there.

Since that implementation is going away, change the one in zmisc.c to
behave the same way.  It used to return a value based on the time since the
process started, which would be variable depending on the implementation.

Both implementations increment based on time spent processing.

This change causes vasarely.ps to run the same with and without dps.