PDF interpreter - preserve stroke colour/space round SMask
Bug #699179 " Chart lines rendered as black when converting from pdf to png or jpg"

Michael ran the problem down to the fact that the stroke colour space
is set before an ExtGState with a Group which has an SMask. The colour
space before the SMask is an ICCBased space and afterwards it is
DeviceGray.

The SMask itself has a Group and that group is defined by a Form XObject
which has (yet another) Group. This Group is defined as having a
DeviceGray space. The problem is that when we executed a SMask in the
interpreter, we saved the fill colour space and values and restored
them at the end (we can't simply use gsave/grestore), but we did not
preserve the stroke colour space and colourants.

This commit saves that as well. This results in a progression with
pdfwrite with one file and three other files showing subtle progressions
with all devices (586_-_missing_images_gs_SMask_not_applied.pdf,
fireworks_with_radial_shading.pdf, x_-_recursive_xobjects_cause_crash.pdf)