pdfwrite - improve h/vmtx table generation for embedded TrueType fonts
Bug 700099 " glyph spacing changed after going through pdfwrite"

This is a difficult one to explain. The primary problem is that the
original PostScript contains multiple EPS files, which contain multiple
embedded subsets of the same font, but all with the same name.

Because these subsets are small, they appear to pdfwrite to be the same
font, the checks we use to check the font do not test every possible
characteristic, because that would be slow.

Now this is not a problem, except that for bug #697376 we did an ugly
hack.

The generation of embedded TrueType fonts badly needs rewritten. Currently
we copy some tables, copy and amend other tables, and generate some tables
from scratch. This has led to numerous problems, on and off.

In the case of bug 697376, we were only writing the hmtx table with
exactly as many metrics as the actual glyphs we put in the output file,
and patched the hhea tale so that it had the correct number of
hLongMetrics entries. But the spec says that every glyph (even those
which are empty in the glyf table!) must have a lsb entry. So Acrobat
Distiller threw an error on encountering our font.

This was solved by copying the hmtx table instead of generating it. Fine
as far as it goes, but in this case we have hit the problem where we are
using glyphs from two different fonts. The hmtx table from the first font
has the wrong metrics for the glyphs in the second font, which leads to
spacing errors.

The only possible solution, until we can find time to rewrite all of this
is to 'pad out' the hmtx table with lsb entries for the glyphs which are
present in the glyf table, but not actually used.

I have a sneaky suspicion that this is still incorrect, we should be
writing the metrics at the correct place (GID) in the table, instead
of as a block, but this works well enough for now.