XPS interpreter - remove special case for artificial bold with pdfwrite
Bug #700382 "Synthetic bold generation too heav"

There was special case code for handling artificially emboldended text
in xps_flush_text_buffer, when the output device could directly handle
text rendering mode 2 (fill and stroke). This was because the pdfwrite
device, the only device capable of handling this mode, did some nasty
trickery with the CTM.

At some point this trickery has been removed, probably as a result of
the ongoing work to handle stroke+fill as a graphics primitive. The
result is that the hackery is no longer required, and in fact causes
the stroke width to be incorrectly calculated, because we are trying to
reverse a later calculation with the CTM, which is no longer applied.

So simply remove the whole section of code. This shows a very few
differences in the test suite, all at low resolution (72 dpi) where
text is slightly bolder, these appear to be (very minor)
progressions.