Bug 700383:  Add support for scRGB
XPS files can define RGB colors with things like

Fill="sc#1, 0.0231533665, 0.08021982, 0.226965874"

This implies that the color space is not sRGB (the standard default in XPS)
but rather scRGB which shares the primaries of sRGB but has a gamma of 1.0.

Lots of progressions in the XPS files expected with this commit.