Bug 698828:  Change of TRC in default_gray.icc
Reviewing this bug revealed that out default_gray.icc profile had a gamma
of 1.8.  The proper gamma should be the same as that used by default_rgb.icc
which has the sRGB color space TRC.  This commit fixes the issue.  Unfortunately
this is going to cause a huge number of changed files.