Bug 700359: Uniprint device did not set color_info.polarity needed by pdf14
The pdf14 device assumes that if the polarity is ADDITIVE, the colorspace
must be Gray or RGB. The uniprint device did not set the polarity when the
colorspace was changed to CMYK, so the number of components set in the pdf14
compositor was 3, even though the ICC profile for the device expects 4 as
num_components in. This caused a mismatch when reading the clist because
it expected 4 components and only 3 were written. Note that the RGBW, RGBOV,
and RGBNOV modes will probably still fail, but unti we see a case for a device
needing these unusual modes, just leave it as it has been for many years.

Note that other devices with strange colorspaces, such as RGBA may also
trip over this.