Bug 699688: Saving pointer into arg_list string invalid (next arg overwrites)
The 'saved_pages_inital_arg' cannot be a pointer into the argument buffer
(arg_list.cstr) since later long arguments can overwrite past the pointer
saved (12 characters into the buffer). Instead allocate a buffer for the
argument and copy the argument into it.

Thanks to Peter Cherepanov for the analysis and part of the fix.