Bug 688210:  Add psdcmyk16 and psdrgb16 devices
So the patch that was provided was far from sufficient in making this
work.  There was significant work that needed to be done in the
pdf14 device to ensure that the transparency put_image operation
worked correctly.  Not sure if the author was paid a bounty but I
question if it was properly tested.

I tested this with some extreme altona files and others and
everything seems to be working well now.  Having the psdcmyk16 device
is nice as it is our only device to have full spot color support and
16 bit output precision.