Add -I handling to gpdl.
This requires an "add_path" entrypoint be added for each language
(NULL for everything except PS).

This causes the block of memory used for the -I path to be leaked. I
suspect this is the same as in gs. A fix for this is a little involved
so it'll wait for now.