Bug 700438: Fix SEGV due to overflowing double->int conversion.
Thanks to Ken for doing most of the investigation on this.

The problem is that we have a double value of 5e9 that won't fit
into an int (and becomes INT_MIN). This huge negative value is then
added to a pointer and we access out of range.

The fix is simply to check for our double values being representable
as ints before we proceed - if not we throw a rangecheck. This is
reasonable as it corresponds to an image being positioned stupidly
far off the page.

The complication with this commit comes in that it collides with
some very suspect old fixes for Bugs 686843 and 687411. The patch
for 686843 makes the same change in 2 places in the code. It looks
like 687411 is reported as the same issue, but the fix for it appears
to be reverting one of the fixes for 686843 - I suspect that the
patch was reversed.

Given that we can't reproduce the original problems that these bugs
were trying to solve any more, I am reverting the code to a saner
form. If the problems recurr, we can reexamine then.