Avoid leak of lib path blocks.
GS keeps a lib_path postscript array with a list of directories
in. Currently this memory is assumed to be "permanent", and is
never freed by the system. This means that whenever we use -I
we leak a block.

Here, we change the system over so that we copy the different
path components into memory blocks and free them at the end.

This seemingly simple task is made much harder by the way that
gs throws out some of the entries in the libpath at random
points during execution.

Accordingly, I have rejigged the code slightly, and introduced
comments to explain what is going on, in the hopes that the
next person through this code (in another 20 years) will have
an easier time understanding the assumptions baked into it.