Rewrite some code to be able to undef internal .setpagesize
/.setpagesize is used internally by gs_statd.ps, but it was written
to pull the current implementation out of statusdict, so it could use
either a level 2-compatible version that called setpagedevice, or a level 1
version that didn't call setpagedevice.

Since we don't support level 1 postscript anymore, we should always have
setpagedevice, so this can be cleaned up.

Now /.setpagesize is just an internal convenience function that we undef
within gs_statd.ps.