gpdl: Add -c operation to gpdl.
We gather up all the commands specified by -c and pass them into
the language as a 'prefix' command, to be run just after we have
inited the job. This means they run inside the encapsulation.

The only language that supports prefix commands is postscript,
hence only postscript commands can be used. This is what we want.