gpdl: Tweak stdin operation to respect gsapi set stdio functions.
Now that gpdl lives entirely under the gsapi API, it should
respect all that API. That means that as people can define
callbacks to be used for stdin/stdout/stderr, we should use
those callbacks if they are set and the program is told to
read from stdin.

Sadly, the only stdin stream implementation we have that supports
this mechanism is the one in the postscript interpreter rather
than the one in the general library, so we can't rely on it being
there (and even if we could, it assumes the presence of an
active PS interpreter instance).

We therefore lift this code out and simplify it to give us a
"callback capable" stdin stream that will be present in all our
builds. This is NOT available using the standard iodev mechanism
so no changes in behaviour can happen for existing programs.