When dereferencing an indirect reference, if we find that the requested object is not in the xref table, return a null object unless PDFSTOPONERROR is set.
When reading an object from the file, verify that the object we read does
in fact have the correct object number. This is because a file can
have an xref which points to a different object, or to no object at all
but a comment. We would then read the next object in the file (if any).