Add code to iterate through Resources looking for an 'real' transparency (Blend Mode not Normal or Compatible, CA or ca not 1.0 any Group outside the page or Annot dictionary)
Add code to check annotations too.

We don't actually use this yet, but we need it so that we can decide
whether any given page needs to push the PDF14 compositor device or
not. We don't want to push it if we don't need it because it has
a negative impact on performance.

Note that there can still be cases where we end up deciding a page is
transparent, when in practice it isn't, such as a soft mask where every
sample is 0xFF, but there's a balance to be struck because the more
checking we do, the longer it takes which reduces the benefit of not
pushing the transparency compositor.