Fix for bug 700042 -- Sanitize out-of-bounds read with fuzzed PDF file.
The test file for this bug was fuzzed in such a way as to generate
Lattice-Form Gouraud shadings with extremely large X coordinates.
This exposed a bug in the code which performs recursive triangle
subdivision.  A color cache entry is not the same size as a
patch_color_t so can't be copied directly.