Add a new PDFINFO switch to dump info ala pdf_info.ps
This isn't completely the same as the pdf_info.ps too from Ghostscript,
it doesn't yet have all the same features.

Currently if you set -dPDFINFO on the command line then the application
will print out the same metadata as pdf_info.ps from the Info dictionary
as well as a count of the number of pages in the file. For every page
it will then emit the various Boxes, whether the page uses transparency
and whether it has any Annotations.

This is an interim piece of work, we'll need it in the long run as we
need to be able to duplicate the functionality of pdf_info.ps so we
may as well have it now. The intended purpose is to test the code which
examines PDF files for transparency by having it produce the same
output as pdf_info.ps, and then comparing the Ghostscript and gpdf
output.