Bug 700482: Binary Object scanner could write past allocated ref array
The object length (osize) was used to compute the max_array_index, but
this exceeded the number of refs allocated based on the overall length
given in the BOS header. We only allocated overall_length/8 but the bad
object length value (in this case for a dict, but could be an array)
was larger so we ended up with a max_array_index that was too large.