Improve -Z: output messages at startup.
The "Init done" messages given by -Z: were misleading.

Here, we simplify the process a bit. We move the init messages
(other than the dumping of the args used) out of -Z: and into
--debug=init-details.

The 3 different init messages are now explicitly done around
each of the 3 phases, rather than in the code that calls
those 3 phases. This makes it far clearer, as the phases can
be called from many different places depending on the exact
command line in use.

The only downside to this is that only the final phase shows up
in normal use, as the parsing of --debug=init_details doesn't
happen until after after the first 2 phases of init have taken
place. This is no worse than the current code, so we'll take this
for now, and update it later (maybe by prescanning the arg list)
if it's ever an issue later on.