New valgrind makefile targets and clist tweak.
We introduce 2 new families of targets to the makefile.

 1) vg, gsvg, gpcl6vg, gxpsvg, gpdlvg
 2) debugvg, gsdebugvg, gpcl6debugvg, gxpsdebugvg, gpdldebugvg

Those build release and debug versions of the executables, tweaked
to work better with valgrind. These mirror the existing *mementovg
targets that were introduced a while ago.

2 extra tweaks are introduced here too.

Firstly, *for these targets*, we default to using memory for
bandlist storage. This should cut down on the number of false
positives we get when using the clist as uninitialised padding
bytes will not be flagged up as they are written to the clist
file. (Instead, they will only be flagged up if they are actually
used during clist playback).

Secondly, whenever we write out a clist block and reset the buffer,
we reset the buffer to "undefined" state so that valgrind can give
us useful information again.