Valgrind: tweak sample_unpack functions to avoid false positive UMRs.
The "sample_unpack" template functions rely on unpacking whole
bytes at a time. Sometimes, the ends of these buffers of bytes
may have bits that are (intentionally, and permissibly) undefined.

Update the code so that in PACIFY_VALGRIND builds we cope with
these undefined bits, and carry the undefinedness over into the
output.