Bug 693747: Fix valgrind problem.
In the image scaling code, we use interpolate_scaled_expanded_height
to calculate a value that will only be used if abs_interp_limit > 1.

The setup code only bothers to initialise the values used for this
calculation if abs_interp_limit > 1, hence valgrind complains that
this is using undefined values.

The simple fix is to only calculate the values if they might be
used.