Fix SEGV found with broken PDF interpreter and Bug690770.pdf
Found accidentally when I was working on a PDF interpreter change, when
closing the pdf14 device, the smaskcolor->ref_count was decremented, but
that object had already been freed (stale pointer). The free was by a
"restore", so the smaskcolor needs to be in stable_memory so that it
will be protected from the restore. Similarly for the "profiles" of that
structure (and free both with stable_memory).