Bug 693784: Tweak copy_mono code for valgrind.
We are smart enough to know that doing a+b is the same as
doing a|b when (a&b)=0 (ignoring overflow). Valgrind is not.

Accordingly our "bit fetching and skewing" code, when run
under valgrind can introduce spurious "undefined" status
to bits when it should not.

The fix is just to use | when we really mean |.

This makes:

  debugbin/gs -sOutputFile=out.pkm -dMaxBitmap=400000000
     -sDEVICE=pkmraw -r300 -sDEFAULTPAPERSIZE=letter
     -dNOPAUSE -dBATCH -K1000000 -dClusterJob -dJOBSERVER
     tests/Ghent_V3.0/010_CMYK_OP_x3.pdf

run to completion with no valgrind warnings.