Fix buffer overread in cmd_compress_bitmap.
The following commands:

 make debugvg
 debugbin/gs -sOutputFile=/dev/null -dMaxBitmap=400000000 -sDEVICE=ppmraw
     -r300 -Z: -dNOPAUSE -dBATCH -K2000000 -dClusterJob out2.ps

cause a read of uninitialised data in cmd_compress_bitmap.

Painstaking debugging shows that the problem is due to an overread
at the end of a bitmap line.

cmd_put_bits is called with 408 bits of bitmap data. This passes the
details into clist_bitmap_bytes, and is told that uncompressed_raster
should be 448 bits.

It then passes this side into cmd_compress_bitmap, which causes the
buffer overread.

The interesting point is that this call is:

    status = cmd_compress_bitmap((stream_state *) & sstate, data,
                          uncompressed_raster << 3 /*width_bits */ ,
                                 raster, height, &w);

which suggests that at some point in the past we were passing the
correct value of width_bits in, and we then changed to passing
the wrong width in.

Sadly, there is no hint in the commit record as to WHY this changed.
This line dates back to the earliest commit we have in source control
in 1998, so it looks like it hasn't changed for a while.

This is only the first part of the problem, though; once in
cmd_compress_bitmap, the value is fed into bitmap_raster() which again
rounds it up so it reads off the end of the defined data.

We therefore rework cmd_compress_bitmap. We specifically only compress
the actual bits of data that we have, and ensure that any padding bits
(or bytes) are correctly sent as constant zeros.

As well as assugaging valgrind, and preventing potential genuine
(if unlikely) buffer overreads, this ensures that we should get
repeatable results on each successive run.