pdfwrite - sort out a type confusion
Bug #700490 "gs_text_enum_t used as pdf_text_enum_t"

There seems to have been 2 sources of confusion here; this falls into
the 'how did this ever work' category.

The problem only occurs when we are forced to go to a bitmap font, which
for PostScript and PDF input is very rare, but for PCL there are several
font types we cannot support and so we must make bitmapped fonts for
them.

We use a pointer in the pdfwrite device structure to store a pointer to
the text enumerator, but we store the *wrong* enumerator. We store
pte_default, which is the enumerator we just created for the standard
test processing. We don't need to store that one, its available as the
argument to all the procedures. What we need access to is the original
text enumerator, the pdfwrite one, and its that one we should be storing
in the device structure.

There was similar confusion in pdf_text_set_cache, which set penum to
'pte' casting it to a pdf_text_enum * in the process. This is clearly
wrong in the case of making a bitmap font, because we made a standard
text enumerator and passed that in, so 'pte' won't be a pdf text
enumerator.

Now it is possible to exercise a different code path and arrive at
pdf_text_set_cache with a pdf_text_enumerator. There is no other way
to arrive here.

So we check pdev->pte (where we would have stored the original pdf
text enumerator) and if its not NULL we use it as a pdf text enumerator.
If it is NULL, then we believe the enumerator pte should he a pdf
text enumerator and cast it as such. Also add a check to ensure that
the enumerator from the arguments *is* a pdf_text enumerator to avoid
this confusion in future.

Finally, clean up a couple of minor compiler warnings in passing.