Bug 694247: Do not confuse openjpeg input buffer usage with size.
Previously s_opjd_accumulate_input() populated the openjpeg input data
buffer and adjusted the buffer size and usage fields accordingly. The
openjpeg callbacks for reading, skipping and seeking confused these two
fields and used the buffer size instead of the buffer usage. This meant
that there was a risk of reading uninitialized data.