pdfwrite - detecting duplicate images, check for pending masks
Bug #700495 "gs/gpdl difference with pdfwrite"

This exhibits with gpdl but not gs because, for some unknown reason, the
MD5 sums from gpdl differ in a few cases from those returned by GS,
I have no idea why.

But that's not the problem, so I'm ignoring it for now.

The real problem is that when we have an image which has an associated
Mask or SMask, when checking to see if it is a duplicate of a prior
image, we were not checking to see if the prior image also had a Mask
or SMask (as appropriate) which could lead to us emitting an image
with a missing mask.

We cater for that now in smask_image_check. Also not strictly part of
this problem, when we did add a Mask or SMask to a stream cos object, we
didn't invalidate its MD5, which could mean it would match a later image
when it should not. So set the md5_valid flag to 0 when we add a mask
entry to the cos object.

On the cluster this show progressions with two files, image-qa.ps and
692217.pdf (drop shadows were previously missing)

Note that the previous output for image-qa.ps was incorrect with both
gs *and* gpdl.