txtwrite - forcibly disable interpolation
Bug #700525 "/unknownerror in --imagemask-- on txtwrite device"

If we have interpolation on, and we get an imagemask, then the rendering
code tries to use get_bits() to perform the interpolation. This
obviously doesn't work with a non-rendering device.

The simplest solution is to disable interpolation, which isn't in any
even useful on a non-rendering device.