pdfwrite: Avoid valgrind errors when comparing image lines.
Within pdf_dump_converted_image we compare scanlines of image data
to the next one. Unfortunately, if scanlines aren't an exact number
of bytes wide, we have uninitialised bits at the end, which can a)
cause spurious mismatches, and b) cause valgrind errors, such as
seen with the following:

 make vg
 valgrind --track-origins=yes bin/gs -sOutputFile=out.ps
    -dMaxBitmap=400000000 -sDEVICE=ps2write -r300 -Z: -dNOPAUSE
    -dBATCH -K2000000 -dClusterJob
    /home/marcos/cluster/tests_private/comparefiles/Bug690237.pdf

Here we simply use a modified line comparison routine to test
the exact set of bits required.