Fix strange valgrind problem; compiler error?
On peeves, the following commands produce valgrind warnings:

 make vg
 bin/gs -sOutputFile=out.ps -dMaxBitmap=400000000 -sDEVICE=ps2write
     -r300 -Z: -dNOPAUSE -dBATCH -K2000000 -dClusterJob
     /home/marcos/cluster/tests_private/comparefiles/Bug691112.pdf
 valgrind --track-origins=yes bin/gs -sOutputFile=/dev/null
     -dMaxBitmap=400000000 -sDEVICE=ppmraw -r300 -Z: -dNOPAUSE
     -dBATCH -K2000000 -dClusterJob out.ps

Debugging shows that the "run_len" member of the RLE stream state
is not being initialised. As far as I can tell, it really out to
be.

Changing the code to use the (neater, IMHO) mechanism of initing
the stream state by calling the template init function seems to
solve this. I cannot for the life of me see why this should make
a difference, but this is a better way to work in any case.